# docs/swagger.json
{
	"components": {
...
		"schemas": {
			"HTTPError": {
				"properties": {
					"code": {
						"type": "number",
						"format": "double"
					},
					"details": {
						"type": "string"
					},
					"additional_info": {}
				},
				"required": [
					"code",
					"details"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Result": {
				"properties": {
					"result": {},
					"error": {
						"$ref": "#/components/schemas/HTTPError"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"errorMsg": {
				"properties": {
					"message": {
						"type": "string"
					},
					"additional_info": {}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IEvent": {
				"properties": {
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"startedAt": {
						"type": "string"
					},
					"endedAt": {
						"type": "string"
					},
					"communityId": {
						"type": "number",
						"format": "double"
					},
					"isPublic": {
						"type": "boolean"
					},
					"imageId": {
						"type": "number",
						"format": "double"
					},
					"type": {
						"type": "string"
					},
					"site": {
						"type": "string"
					},
					"capacity": {
						"type": "number",
						"format": "double"
					},
					"additionalInfo": {
						"type": "string"
					}
				},
				"required": [
					"title",
					"startedAt",
					"isPublic",
					"type",
					"site",
					"capacity"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"accessToken": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "backend",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/communities/{CID}/events": {
			"post": {
				"operationId": "CreateEvent",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Result"
								},
								"examples": {
									"Example 1": {
										"value": {
											"data": {
												"id": 1
											}
										}
									}
								}
							}
						}
					},
					"403": {
						"description": "Unathorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/errorMsg"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "You are not an admin of this community"
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/errorMsg"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Community not found"
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/errorMsg"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Failed to get community"
										}
									}
								}
							}
						}
					}
				},
				"description": "Create events in a community",
				"tags": [
					"Community"
				],
				"security": [],
				"parameters": [
					{
						"description": "Community ID",
						"in": "path",
						"name": "CID",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IEvent"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetEvents",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Result"
								},
								"examples": {
									"Example 1": {
										"value": {
											"data": {
												"id": 1,
												"name": "Komunitas IT",
												"slug": null,
												"description": "Komunitas IT",
												"createdAt": "2022-04-01T20:07:48.687Z",
												"updatedAt": "2022-04-01T20:07:48.687Z",
												"events": []
											}
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/errorMsg"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Community not found"
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/errorMsg"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Failed to get community"
										}
									}
								}
							}
						}
					}
				},
				"description": "Get events in a community",
				"tags": [
					"Community"
				],
				"security": [],
				"parameters": [
					{
						"description": "Community ID",
						"in": "path",
						"name": "CID",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "true to get upcoming events or false to get past events",
						"in": "query",
						"name": "upcoming",
						"required": true,
						"schema": {
							"type": "boolean"
						}
					}
				]
			}
		},
...
		"/communities/{CID}/members/status": {
			"get": {
				"operationId": "GetMembershipStatus",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Result"
								},
								"examples": {
									"Example 1": {
										"value": {
											"data": {
												"status": "member"
											}
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/errorMsg"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "User not authorized"
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/errorMsg"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Community not found"
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/errorMsg"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Failed to get community"
										}
									}
								}
							}
						}
					}
				},
				"description": "Get user's membership status in a community",
				"tags": [
					"Community"
				],
				"security": [
					{
						"accessToken": []
					}
				],
				"parameters": [
					{
						"description": "id of the community",
						"in": "path",
						"name": "CID",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/events": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Result"
								},
								"examples": {
									"Example 1": {
										"value": {
											"data": {
												"events": [
													{
														"id": 1,
														"name": "Event 1",
														"description": "Event 1 description",
														"startDate": "2020-01-01",
														"endDate": "2020-01-01",
														"communityId": 1,
														"participants": [
															{
																"id": 1,
																"name": "John Doe",
																"email": "johndoa@doe.com",
																"phone": "123456789",
																"isAdmin": false
															}
														]
													}
												]
											}
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/errorMsg"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Could not get events"
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"Event"
				],
				"security": [],
				"parameters": []
			}
		},
		"/events/{EID}/join": {
			"post": {
				"operationId": "Join",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Result"
								},
								"examples": {
									"Example 1": {
										"value": {}
									}
								}
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/errorMsg"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Event is full"
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "community not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/errorMsg"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Community not found"
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"Event"
				],
				"security": [
					{
						"accessToken": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "EID",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}